Improved math expressions
Math expressions can now handle precedence and parenthesis. ID reference is no longer valid mathExpression.
Added the notion of a mathFunction to prepare for build-in mathematics.
Added behind the scene float and ints, so we cant iterate 1.5 times in a loop.

Added helper non-terminals. Added  assignment rule, but it is currently not in use.