Deleted some duplicate nodes
Made the decision that ShapeNodes will be the same as shape declaration nodes. Therefore added a body to them.
Added lots of nodetypes for the AST.
Only need to be able to build BodyNodes from BodtContexts, then we have a good tree. Will implement Visitor for this in next commit