Ast builder visitor (#21)
* CreatedAST builder

Made a container responsible for semantic analysis.
Made  lots of nodes for the AST. Now we need to decide how to build em and structure em with properties

* Deleted some duplicate nodes

Made the decision that ShapeNodes will be the same as shape declaration nodes. Therefore added a body to them.
Added lots of nodetypes for the AST.
Only need to be able to build BodyNodes from BodtContexts, then we have a good tree. Will implement Visitor for this in next commit

* Created ErrorInheritorVisitor for ease of passing error lists

* Lots of extracting nodes

Working on extracting assignments. Currently, it is possible to create MathNodes, but we want  a more specific tree (AdditionNodes for instance).
Same for point references.

* Working on extracting math expressions

* MathNodes and bool nodes are now almost complete.

We can now create bool nodes and math nodes. These are created using Extractors. ExtractNode methods are used when one have to typecast from one context to other context to find the actual context type.

TODO: Create MathFunctionCallNodes
 and BoolFunctionCallNodes and extractors so we have nodes  that can be represented in math and bool exprs

* Small commit

* Lots of stuff

Co-authored-by: Rasmus Hoeyer Hansen <rasmus.hoeyer.hansen@gmail.com>